MAINT: renaming files and directories
