--- enigma-principal/app/src/boot_network/mod.rs
@@ -1,2 +1,0 @@
-pub mod km_http_server;
-pub mod principal_manager;
