--- enigma-principal/app/src/cli/mod.rs
@@ -1,2 +1,0 @@
-pub mod app;
-pub mod options;
