+++ enigma-principal/app/src/controller/config.json
