+++ enigma-principal/app/src/controller/mod.rs
@@ -1,4 +1,5 @@
 pub mod km_controller;
-pub mod epoch_types;
-pub mod verifier;
 pub mod km_utils;
+pub mod boot_network;
+pub mod options;
+pub mod km_http_server;
