+++ enigma-principal/app/src/epochs/epoch_keeper_u.rs
@@ -1,12 +1,13 @@
-use enigma_tools_m::keeper_types::InputWorkerParams;
 use failure::Error;
 use rustc_hex::ToHex;
 use sgx_types::{sgx_enclave_id_t, sgx_status_t};
 use web3::types::{Bytes, U256};
 
-use common_u::errors::EnclaveFailError;
+use enigma_tools_m::keeper_types::InputWorkerParams;
 use enigma_types::{EnclaveReturn, traits::SliceCPtr};
-use controller::epoch_types::{encode, SignedEpoch};
+
+use common_u::errors::EnclaveFailError;
+use epochs::epoch_types::{encode, SignedEpoch};
 
 extern "C" {
     fn ecall_set_worker_params(
