+++ enigma-principal/app/src/epochs/epoch_types.rs
@@ -1,14 +1,15 @@
 use std::collections::HashMap;
-use rustc_hex::ToHex;
 
-use enigma_tools_m::keeper_types::InputWorkerParams;
+use rustc_hex::ToHex;
 use ethabi::{Event, EventParam, ParamType};
 use failure::Error;
 pub use rlp::{decode, Encodable, encode, RlpStream};
 use serde::{Deserialize, Serialize};
 use web3::types::{Address, Bytes, H160, U256};
 
+use enigma_tools_m::keeper_types::InputWorkerParams;
 use enigma_types::Hash256;
+
 use common_u::errors::EpochStateTransitionErr;
 
 pub const EPOCH_STATE_UNCONFIRMED: &str = "UNCONFIRMED";
