+++ enigma-principal/app/src/esgx/mod.rs
@@ -1,4 +1,2 @@
-pub mod epoch_keeper_u;
 pub mod equote;
 pub mod general;
-pub mod keys_keeper_u;
