+++ enigma-principal/app/src/state_keys/keys_keeper_u.rs
@@ -4,9 +4,10 @@ use failure::Error;
 use sgx_types::{sgx_enclave_id_t, sgx_status_t};
 use web3::types::U256;
 
-use boot_network::km_http_server::{StateKeyRequest, StateKeyResponse};
-use common_u::errors::EnclaveFailError;
 use enigma_types::{Hash256, EnclaveReturn, traits::SliceCPtr};
+
+use controller::km_http_server::{StateKeyRequest, StateKeyResponse};
+use common_u::errors::EnclaveFailError;
 
 extern "C" {
     fn ecall_get_enc_state_keys(
@@ -63,13 +64,14 @@ pub fn get_enc_state_keys(eid: sgx_enclave_id_t, request: StateKeyRequest, epoch
 
 #[cfg(test)]
 pub mod tests {
-    use sgx_urts::SgxEnclave;
-
-    use esgx::epoch_keeper_u::set_or_verify_worker_params;
-    use esgx::epoch_keeper_u::tests::get_worker_params;
-    use esgx::general::init_enclave_wrapper;
-    use rustc_hex::FromHex;
     use super::*;
+
+    use sgx_urts::SgxEnclave;
+    use rustc_hex::FromHex;
+
+    use epochs::epoch_keeper_u::set_or_verify_worker_params;
+    use epochs::epoch_keeper_u::tests::get_worker_params;
+    use esgx::general::init_enclave_wrapper;
 
     fn init_enclave() -> SgxEnclave {
         let enclave = match init_enclave_wrapper() {
